// frontend/src/app/openapi/model/project-user.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { User } from './user';
import { ProjectUserRole } from './project-user-role';
import { ProjectUserPermission } from './project-user-permission';


export interface ProjectUser { 
    role: ProjectUserRole;
    permission: ProjectUserPermission;
    user: User;
}
export namespace ProjectUser {
}